requirements.txt
langchain==0.3.24
langchain_cohere==0.4.4
langchain_community==0.3.22
langchain_core==0.3.56
langchain_mistralai==0.2.10
langchain_qdrant==0.2.0
neo4j==5.25.0
pydantic==2.11.3
qdrant_client==1.14.2
Requests==2.32.3
streamlit==1.43.2
streamlit_reveal_slides==0.2.0
